updated to be closer to intentions
